server/requirements.txt
Flask==2.0.2
python-dotenv==0.19.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
Flask-JWT-Extended==4.3.1
numpy==1.22.1
pandas==1.4.0
scikit-learn==0.24.1